requirements.txt
streamlit
pillow
openai
requests